+++ src/app/shared/validators.ts
@@ -1,0 +1,25 @@
+import { AbstractControl, ValidatorFn } from '@angular/forms';
+
+export function UniqueValidator(uniqueList: any[]): ValidatorFn {
+    return (control: AbstractControl): {[key: string]: any} | null => {
+        if (uniqueList.indexOf(control.value) !== -1) {
+            return {
+                nonUnique: {
+                    value: control.value,
+                },
+            };
+        }
+    };
+}
+
+export function ListValidator(allowList: any[]): ValidatorFn {
+    return (control: AbstractControl): {[key: string]: any} | null => {
+        if (allowList.indexOf(control.value) === -1) {
+            return {
+                notInList: {
+                    value: control.value,
+                },
+            };
+        }
+    };
+}