+++ src/app/transaction-form/account-modal/account-modal.component.ts
@@ -1,8 +1,11 @@
 import { Component, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
-
+import { FormControl } from '@angular/forms';
 import { ModalDialogParams } from 'nativescript-angular/modal-dialog';
 
+import { ListPicker } from 'tns-core-modules/ui/list-picker';
+
 import { showKeyboard } from '../../shared/misc';
+import { ListValidator } from '../../shared/validators';
 
 @Component({
     selector: 'bc-account-modal',
@@ -14,12 +17,15 @@ export class AccountModalComponent implements AfterViewInit {
     @ViewChild('accountField', {static: false})
     accountField: ElementRef;
 
+    account: FormControl;
     accounts: string[];
-    search: string;
     selectedIndex: number;
 
     constructor(private modalParams: ModalDialogParams) {
         this.accounts = modalParams.context;
+        this.account = new FormControl('', [
+            ListValidator(this.accounts),
+        ]);
     }
 
     ngAfterViewInit(): void {
@@ -27,19 +33,33 @@ export class AccountModalComponent implements AfterViewInit {
     }
 
     filterAccounts(): void {
-        const regexp = new RegExp(this.search, 'iu');
+        const regexp = new RegExp(this.account.value, 'iu');
         this.accounts = this.modalParams.context.filter((account) => {
             return account.search(regexp) !== -1;
         });
     }
 
-    onCancel(): void {
+    onAccountTap(args): void {
+        const picker = <ListPicker>args.object;
+        if (picker.selectedIndex === -1) {
+            // Empty list
+            return;
+        }
+        const selectedValue = this.accounts[picker.selectedIndex];
+        if (selectedValue === this.account.value) {
+            // Item already selected, close modal on second tap
+            this.select();
+        } else {
+            this.account.setValue(selectedValue);
+        }
+    }
+
+    cancel(): void {
         this.modalParams.closeCallback(null);
     }
 
-    onSelect(): void {
-        const account = this.accounts[this.selectedIndex];
-        this.modalParams.closeCallback(account);
+    select(): void {
+        this.modalParams.closeCallback(this.account.value);
     }
 
 }